requirements.txt
Flask
Flask-SQLAlchemy
werkzeug
Flask-jwt-extended 
Flask-Bcrypt
flask_cors
requests
schedule
pandas
paho_mqtt